implementing long string sending
this is so I can send the long string of data to the chassis relating to where pixels are going to be during the movement. needs to be complicated because espnow limits data packet size to 250 bytes/characters. this is untested gonna try to test it in a few hours